+++ documentation/changelog.md
@@ -14,6 +14,7 @@ description: List of all the changes and current versions.
 + Add Basic Unit Tests
 + Add Effect Based Items
 + Fixed bug when pressing 1-4 on keyboard while typing it would pull items.
++ Added /getitem Command
 ```
 
 ## 1.3.0 - January 23-26, 2021